+++ flutter_secure_storage_platform_interface/lib/flutter_secure_storage_platform_interface.dart
@@ -1,6 +1,6 @@
 library flutter_secure_storage_platform_interface;
 
-import 'dart:io';
+// import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
@@ -31,9 +31,9 @@ abstract class FlutterSecureStoragePlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Stream<bool> get onCupertinoProtectedDataAvailabilityChanged;
+  // Stream<bool> get onCupertinoProtectedDataAvailabilityChanged;
 
-  Future<bool> isCupertinoProtectedDataAvailable();
+  // Future<bool> isCupertinoProtectedDataAvailable();
 
   Future<void> write({
     required String key,